+++ src/core_modules/capture-core/components/WidgetEnrollmentEventNew/DataEntry/epics/dataEntryRules.epics.js
@@ -1,6 +1,7 @@
 // @flow
 import { ofType } from 'redux-observable';
-import { map } from 'rxjs/operators';
+import { map, switchMap } from 'rxjs/operators';
+import { from } from 'rxjs';
 import { batchActions } from 'redux-batched-actions';
 import type { OrgUnit } from '@dhis2/rules-engine-javascript';
 import i18n from '@dhis2/d2-i18n';
@@ -15,22 +16,34 @@ import {
     getCurrentClientMainData,
     getApplicableRuleEffectsForTrackerProgram,
     updateRulesEffects,
+    validateAssignEffects,
     type FieldData,
 } from '../../../../rules';
 import { getDataEntryKey } from '../../../DataEntry/common/getDataEntryKey';
 import type { RulesExecutionDependenciesClientFormatted } from '../../common.types';
 import { getLocationQuery } from '../../../../utils/routing';
+import type { QuerySingleResource } from '../../../../utils/api';
 
-const runRulesForNewEvent = (
+const runRulesForNewEvent = async ({
+    store,
+    dataEntryId,
+    itemId,
+    uid,
+    orgUnit,
+    rulesExecutionDependenciesClientFormatted,
+    fieldData,
+    querySingleResource,
+}: {
     store: ReduxStore,
     dataEntryId: string,
     itemId: string,
     uid: string,
     orgUnit: OrgUnit,
-    history: Object,
-    { events, attributeValues, enrollmentData }: RulesExecutionDependenciesClientFormatted,
+    rulesExecutionDependenciesClientFormatted: RulesExecutionDependenciesClientFormatted,
     fieldData?: ?FieldData,
-) => {
+    querySingleResource: QuerySingleResource,
+}) => {
+    const { events, attributeValues, enrollmentData } = rulesExecutionDependenciesClientFormatted;
     const state = store.value;
     const formId = getDataEntryKey(dataEntryId, itemId);
     const { programId, stageId } = getLocationQuery();
@@ -47,7 +60,7 @@ const runRulesForNewEvent = (
     const currentEventMainData = getCurrentClientMainData(state, itemId, dataEntryId, foundation);
     const currentEvent = { ...currentEventValues, ...currentEventMainData, programStageId };
 
-    const ruleEffects = getApplicableRuleEffectsForTrackerProgram({
+    const effects = getApplicableRuleEffectsForTrackerProgram({
         program,
         stage,
         orgUnit,
@@ -57,40 +70,55 @@ const runRulesForNewEvent = (
         enrollmentData,
     });
 
+    const effectsWithValidations = await validateAssignEffects({
+        dataElements: foundation.getElements(),
+        effects,
+        querySingleResource,
+    });
+
     return batchActions([
-        updateRulesEffects(ruleEffects, formId),
+        updateRulesEffects(effectsWithValidations, formId),
         rulesExecutedPostUpdateField(dataEntryId, itemId, uid),
     ],
     newEventWidgetDataEntryBatchActionTypes.RULES_EFFECTS_ACTIONS_BATCH,
     );
 };
 
-export const runRulesOnUpdateDataEntryFieldForNewEnrollmentEventEpic = (action$: InputObservable, store: ReduxStore, { history }: ApiUtils) =>
+export const runRulesOnUpdateDataEntryFieldForNewEnrollmentEventEpic = (
+    action$: InputObservable,
+    store: ReduxStore,
+    { querySingleResource }: ApiUtils,
+) =>
     action$.pipe(
         ofType(newEventWidgetDataEntryBatchActionTypes.UPDATE_DATA_ENTRY_FIELD_ADD_EVENT_ACTION_BATCH),
         map(actionBatch =>
             actionBatch.payload
                 .find(action => action.type === newEventWidgetDataEntryActionTypes.RULES_ON_UPDATE_EXECUTE)),
-        map((action) => {
+        switchMap((action) => {
             const { dataEntryId, itemId, uid, orgUnit, rulesExecutionDependenciesClientFormatted } = action.payload;
-            return runRulesForNewEvent(
+            const runRulesForNewEventPromise = runRulesForNewEvent({
                 store,
                 dataEntryId,
                 itemId,
                 uid,
                 orgUnit,
-                history,
                 rulesExecutionDependenciesClientFormatted,
-            );
+                querySingleResource,
+            });
+            return from(runRulesForNewEventPromise);
         }));
 
-export const runRulesOnUpdateFieldForNewEnrollmentEventEpic = (action$: InputObservable, store: ReduxStore, { history }: ApiUtils) =>
+export const runRulesOnUpdateFieldForNewEnrollmentEventEpic = (
+    action$: InputObservable,
+    store: ReduxStore,
+    { querySingleResource }: ApiUtils,
+) =>
     action$.pipe(
         ofType(newEventWidgetDataEntryBatchActionTypes.FIELD_UPDATE_BATCH),
         map(actionBatch =>
             actionBatch.payload
                 .find(action => action.type === newEventWidgetDataEntryActionTypes.RULES_ON_UPDATE_EXECUTE)),
-        map((action) => {
+        switchMap((action) => {
             const {
                 dataEntryId,
                 itemId,
@@ -107,14 +135,16 @@ export const runRulesOnUpdateFieldForNewEnrollmentEventEpic = (action$: InputObs
                 value,
                 valid: uiState.valid,
             };
-            return runRulesForNewEvent(
+
+            const runRulesForNewEventPromise = runRulesForNewEvent({
                 store,
                 dataEntryId,
                 itemId,
                 uid,
                 orgUnit,
-                history,
                 rulesExecutionDependenciesClientFormatted,
                 fieldData,
-            );
+                querySingleResource,
+            });
+            return from(runRulesForNewEventPromise);
         }));