+++ src/core_modules/capture-core/components/WidgetRelatedStages/WidgetRelatedStages.component.js
@@ -33,6 +33,7 @@ const WidgetRelatedStagesPlain = ({
     const [relatedStageDataValues, setRelatedStageDataValues] = useState<RelatedStageDataValueStates>({
         linkMode: undefined,
         scheduledAt: '',
+        scheduledAtFormatError: undefined,
         orgUnit: undefined,
         linkedEventId: undefined,
     });
@@ -61,10 +62,11 @@ const WidgetRelatedStagesPlain = ({
     };
 
     const formIsValid = useCallback(() => {
-        const { scheduledAt, orgUnit, linkedEventId, linkMode } = relatedStageDataValues;
+        const { scheduledAt, scheduledAtFormatError, orgUnit, linkedEventId, linkMode } = relatedStageDataValues;
         return relatedStageWidgetIsValid({
             linkMode,
             scheduledAt,
+            scheduledAtFormatError,
             orgUnit,
             linkedEventId,
             setErrorMessages: addErrorMessage,