+++ src/core_modules/capture-core/components/WidgetRelatedStages/relatedStageEventIsValid/relatedStageEventIsValid.types.js
@@ -5,6 +5,7 @@ import { RelatedStageModes } from '../index';
 export type RelatedStageIsValidProps = {|
     linkMode: ?$Keys<typeof RelatedStageModes>,
     scheduledAt: ?string,
+    scheduledAtFormatError: ?{error: ?string, errorCode: ?string},
     orgUnit: ?{
         id: string,
         name: string,