+++ src/core_modules/capture-core/utils/validators/form/ageValidator.js
@@ -1,8 +1,0 @@
-// @flow
-import { isValidAge as isValidAgeCore } from 'capture-core-utils/validators/form';
-import { systemSettingsStore } from '../../../metaDataMemoryStores';
-
-export function isValidAge(value: Object) {
-    const format = systemSettingsStore.get().dateFormat;
-    return isValidAgeCore(value, format);
-}